Changed the functions to their own .c files and made .h files. I had some problem on the testing, due to my IDE and Compiler, but it will run... probably
