+++ AureaKx_main.c
@@ -1,25 +1,22 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <string.h>
-#include <math.h>
-#include <tgmath.h>
 
-// -------------------------------------- Structs ----------------------------------------- \\
+// ------------------------------- Intern Libraries --------------------------------- \\
+
+#include "Shufflers.h"
+#include "Solutioners.h"
+
+// ----------------------------------- Structs -------------------------------------- \\
 
 struct hash1024 {
     uint8_t hexBox[128];
     uint16_t id;
 };
 
-// --------------------- Function Prototypes --------------------- \\
+// ------------------------------ Function Prototypes -------------------------------- \\
 
-void converter(const char *input, size_t length, uint8_t *hexBox);
-void autoFill(const uint8_t *hexBox, size_t length, size_t capacity, uint8_t *hashBoxH);
-void compactor (const uint8_t *hexBox, size_t length, size_t capacity, uint8_t *hashBoxH);
-void goldenShuffler(uint8_t *hashBox, size_t capacity, int rounds);
-void eulerShuffler(uint8_t *hashBox, size_t capacity, int rounds);
-void differentiator(uint8_t *hashBox, size_t capacity);
-void corrector(uint8_t *hashBox, size_t capacity, uint8_t *inputOriginal, size_t sizeOriginal);
+
 
 // -------------------------------- Main ----------------------------------- \\
 
@@ -86,143 +83,3 @@ int main() {
 }
 
 // ---------------------------- Functions -------------------------------- \\
-
-void converter(const char *input, size_t length, uint8_t *hexBox) {
-    memcpy(hexBox, input, length);
-}
-
-void autoFill(const uint8_t *hexBox, size_t length, const size_t capacity, uint8_t *hashBoxH) {
-
-    const uint8_t container_ASCII[] = {
-        0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29,
-        0x2A, 0x2B, 0x2C, 0x2D, 0x2E, 0x2F, 0x30, 0x31, 0x32, 0x33,
-        0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x3A, 0x3B, 0x3C, 0x3D,
-        0x3E, 0x3F, 0x40, 0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47,
-        0x48, 0x49, 0x4A, 0x4B, 0x4C, 0x4D, 0x4E, 0x4F, 0x50, 0x51,
-        0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58, 0x59, 0x5A, 0x5B,
-        0x5C, 0x5D, 0x5E, 0x5F, 0x60, 0x61, 0x62, 0x63, 0x64, 0x65,
-        0x66, 0x67, 0x68, 0x69, 0x6A, 0x6B, 0x6C, 0x6D, 0x6E, 0x6F,
-        0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77, 0x78, 0x79,
-        0x7A, 0x7B, 0x7C, 0x7D, 0x7E
-    };
-    int container_size = sizeof(container_ASCII) / sizeof(container_ASCII[0]);
-
-    memcpy(hashBoxH, hexBox, length);
-
-    // semi-Perfection (just doubles one-time the same ASCII):
-    for (size_t j = length; j < capacity; j++) {
-        double planFei_jj = ((6.62606*(j/4.66920)) * j);
-        int pfjjR = round(planFei_jj);
-
-        if (pfjjR < container_size) {
-            int selected = container_ASCII[pfjjR % container_size];
-            hashBoxH[j] = selected;
-        }
-        else {
-            int temp = hashBoxH[(j * pfjjR) % length];
-            int selected = container_ASCII[((temp * pfjjR) % j) % container_size];
-            hashBoxH[j] = selected;
-        }
-    }
-}
-
-void compactor(const uint8_t *hexBox, size_t length, const size_t capacity, uint8_t *hashBoxH) {
-    // Temporary buffer to work on
-    uint8_t tempBox[length];
-    memcpy(tempBox, hexBox, length);
-    size_t currentLength = length;
-
-    // Reducing until 128 items
-    while (currentLength > capacity) {
-        size_t newLength = 0;
-        for (size_t i = 0; i < currentLength; i += 2) {
-            uint16_t sum = tempBox[i];
-            if (i + 1 < currentLength) {
-                sum += tempBox[i + 1];
-            }
-            // moderator
-            uint8_t reduced = (sum > 127) ? (sum % 127) : sum;
-            tempBox[newLength++] = reduced;
-        }
-        currentLength = newLength;
-    }
-
-    // Passing temp to hash
-    memcpy(hashBoxH, tempBox, capacity);
-}
-
-void goldenShuffler(uint8_t *hashBox, const size_t capacity, const int rounds) {
-    const double golden = 1.618033988;
-    uint8_t tempBox[capacity];
-    int roundsG = rounds * round(golden);
-    memcpy(tempBox, hashBox, capacity);
-
-    for (int i = 0; i < roundsG; i++) {
-        double gM = golden * (i*i);
-        int gM_R = round(gM);
-        uint8_t Temp1 = tempBox[gM_R % capacity];
-        tempBox[(i*2) % capacity] = Temp1;
-    }
-
-    memcpy(hashBox, tempBox, capacity);
-}
-
-void eulerShuffler(uint8_t *hashBox, const size_t capacity, const int rounds) {
-    const double euler = 2.71828;
-    uint8_t tempBox[capacity];
-    int roundsE = rounds * (round(euler*5.77));
-    memcpy(tempBox, hashBox, capacity);
-
-    for (int i = 0; i < roundsE; i++) {
-        double eM = euler * (i*i);
-        int eMR = round(eM);
-        uint8_t Temp1 = tempBox[eMR % capacity];
-        tempBox[(i*2) % capacity] = Temp1;
-    }
-
-    memcpy(hashBox, tempBox, capacity);
-}
-
-void differentiator(uint8_t *hashBox, const size_t capacity) {
-    uint8_t tempBox[capacity];
-    memcpy(tempBox, hashBox, capacity);
-
-    for (int i = 0; i < capacity; i++) {
-        int preTemp = tempBox[i];
-        int futTemp = tempBox[i+1];
-
-        if (preTemp == futTemp) {
-            int verTemp = preTemp + 1;
-            tempBox[i] = verTemp;
-        }
-        else {
-            tempBox[i] = preTemp;
-        }
-    }
-
-    memcpy(hashBox, tempBox, capacity);
-}
-
-void corrector(uint8_t *hashBox, const size_t capacity, uint8_t *inputOriginal, const size_t sizeOriginal) {
-    uint8_t tempBox[capacity];
-    memcpy(tempBox, hashBox, capacity);
-
-    // Sum all values
-    uint32_t sumHash = 0;
-    for (size_t i = 0; i < sizeOriginal; i++) {
-        sumHash += inputOriginal[i];
-    }
-
-    // Determine if the sum is odd or even
-    uint8_t correction = (sumHash % 2 == 0) ? 2 : 1;
-
-    // Apply the correction to specified indices
-    size_t indices[] = {1, 3, 5, 7, 31};
-    for (size_t i = 0; i < 5; i++) {
-        size_t idx = indices[i];
-        tempBox[idx] += correction;
-    }
-
-    memcpy(hashBox, tempBox, capacity);
-
-}
