+++ Logics.c
@@ -1,0 +1,5 @@
+//
+// Created by thiag on 06/08/2025.
+//
+
+#include "Logics.h"
