+++ Logics.h
@@ -1,0 +1,8 @@
+//
+// Created by thiag on 06/08/2025.
+//
+
+#ifndef LOGICS_H
+#define LOGICS_H
+
+#endif //LOGICS_H
