+++ Shufflers.h
@@ -1,0 +1,7 @@
+#ifndef SHUFFLERS_H
+#define SHUFFLERS_H
+
+void goldenShuffler(uint8_t *hashBox, size_t capacity, int rounds);
+void eulerShuffler(uint8_t *hashBox, size_t capacity, int rounds);
+
+#endif //SHUFFLERS_H
